requirements.txt
atcoder-tools>=2.12.0
lark>=1.0.0
lxml>=4.7.1
pyperclip>=1.8.2
pytest>=7.1.1
PyYAML>=6.0
selenium>=4.1.0
typing_extensions>=4.0.1
webdriver-manager>=3.5.2
